Fix additional PR comments.
