+++ encoding/math.go
@@ -1,5 +1,7 @@
 package encoding
 
+import "time"
+
 func intSubIntFn(v, v2 int) int {
 	return v - v2
 }
@@ -18,4 +20,12 @@ func int64AddIntFn(v int64, v2 int) int64 {
 
 func int64SubInt64Fn(v, v2 int64) int {
 	return int(v - v2)
+}
+
+func scaleDownFn(v int64, resolution time.Duration) int64 {
+	return v / int64(resolution)
+}
+
+func scaleUpFn(v int64, resolution time.Duration) int64 {
+	return v * int64(resolution)
 }
